+++ TJC.ConsoleApplication.Tests/Options/SpecificArgumentTests.cs
@@ -4,7 +4,6 @@ using TJC.ConsoleApplication.Arguments.Options.Specific;
 
 namespace TJC.ConsoleApplication.Tests.Options;
 
-
 public class SpecificArgumentTests
 {
     [Fact]
@@ -30,6 +29,26 @@ public class SpecificArgumentTests
         arguments.ParseAndValidate(["--verbose=2", "--verbose"], exitOnFailureToParse: false);
 
         Assert.Equal(3, argument.Verbosity);
+    }
+
+    [Theory]
+    [InlineData("-v", 1)]
+    [InlineData("-vv", 2)]
+    [InlineData("-vvv", 3)]
+    [InlineData("-vvvv", 4)]
+    [InlineData("-vvvvv", 5)]
+    [InlineData("-vvvvvv", 6)]
+    [InlineData("-vvvvvvv", 7)]
+    [InlineData("-vvvvvvvv", 8)]
+    public void VerbosityArgument_AccumulatesClusteredShortOptions(string input, int expectedVerbosity)
+    {
+        var argument = VerbosityArgument.Both;
+        argument.Argument.ExitIfUsed = false;
+        var arguments = new ConsoleArguments { argument };
+
+        arguments.ParseAndValidate([input], exitOnFailureToParse: false);
+
+        Assert.Equal(expectedVerbosity, argument.Verbosity);
     }
 
     [Fact]