+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- The verbosity argument now allows multiple v's to set verbosity level, e.g. `-vvvvv` is the same as `-v=5`
+
 ### Fixed
 
 - Fix grype lint errors